On second thought, let's use what I already had...
in my [website repo](https://github.com/rswinkle/rswinkle.github.io).
The goal for all these README.md edits today (in other repos as well) is
to make the READMEs match what I use for generating the project pages for my
static website.  Obviously while having a download link to the github
repo doesn't make sense for someone viewing the project on github, if
they're on my website viewing the html generated from the README...

And for this README, all those convenient relative links to images in
the repo break for my website so they have to be full urls.