Add Scissoring behavior test...
I'm adding programs like this and test_poly_modes to ensure I know what
the behavior *should* be, before I write regression tests for PortableGL
for the same features.  Sometimes the test will be exactly the same (as
I plan to do for scissoring), sometimes I just needed to double check
how the feature works or interacts with other state before coming up
with a test.